Move private message dialogs into addonStoreGui
